+++ src/values/arrays.test-d.ts
@@ -69,6 +69,7 @@ describe("some()", () => {
 
 });
 
+
 describe("unique()", () => {
 
 	it("should infer the element type from any collection", async () => {